Move higher-order function to functools
