+++ configs/csg/pbe_with_repl_pretrain_base.yaml
@@ -10,17 +10,17 @@ optimizer:
     model: "@/model"
 
 collate_fn:
-    type: mlprogram.utils.Sequence
+    type: mlprogram.functools.Sequence
     funcs:
         type: collections.OrderedDict
         items:
             - - "to_episode"
-              - type: mlprogram.utils.Map
+              - type: mlprogram.functools.Map
                 func: "@/to_episode"
             - - "flatten"
               - type: mlprogram.utils.Flatten
             - - "transform"
-              - type: mlprogram.utils.Map
+              - type: mlprogram.functools.Map
                 func: "@/transform"
             - - "collate"
               - "@/collate.collate"
