+++ configs/csg/pbe_with_repl_train_base.yaml
@@ -21,17 +21,17 @@ batch_size: 1
 n_rollout: 16
 
 collate_fn:
-  type: mlprogram.utils.Sequence
+  type: mlprogram.functools.Sequence
   funcs:
     type: collections.OrderedDict
     items:
       - - "to_episode"
-        - type: mlprogram.utils.Map
+        - type: mlprogram.functools.Map
           func: "@/to_episode"
       - - "flatten"
         - type: mlprogram.utils.Flatten
       - - "transform"
-        - type: mlprogram.utils.Map
+        - type: mlprogram.functools.Map
           func: "@/transform"
       - - "collate"
         - "@/collate.collate"
