+++ configs/nl2bash/nl2code_train.yaml
@@ -4,7 +4,7 @@ imports:
 output_dir: "output/output"
 
 transform:
-  type: mlprogram.utils.Sequence
+  type: mlprogram.functools.Sequence
   funcs:
     type: collections.OrderedDict
     items:
@@ -61,12 +61,12 @@ main:
   metric: "bleu@3"
   threshold: 1.0
   collate:
-    type: mlprogram.utils.Compose
+    type: mlprogram.functools.Compose
     funcs:
       type: collections.OrderedDict
       items:
         - - "transform"
-          - type: mlprogram.utils.Map
+          - type: mlprogram.functools.Map
             func: "@/transform"
         - - "collate"
           - "@/collate.collate"
