+++ configs/nl2code/base.yaml
@@ -1,59 +1,3 @@
-device:
-  type: torch.device
-  type_str: "cuda"
-  index: 0
-
-parser:
-  type: mlprogram.datasets.django.Parser
-  split_value:
-    type: mlprogram.datasets.django.SplitValue
-
-dataset:
-  type: mlprogram.datasets.django.download
-train_dataset: "@/dataset.train"
-test_dataset:
-  type: mlprogram.utils.data.transform
-  dataset: "@/dataset.test"
-  transform: "@/normalize_dataset"
-valid_dataset:
-  type: mlprogram.utils.data.transform
-  dataset: "@/dataset.valid"
-  transform: "@/normalize_dataset"
-
-words:
-  type: mlprogram.utils.data.get_words
-  dataset: "@/train_dataset"
-  extract_reference:
-    type: mlprogram.datasets.django.TokenizeQuery
-samples:
-  type: mlprogram.utils.data.get_samples
-  dataset: "@/train_dataset"
-  parser: "@/parser"
-
-encoder:
-  word_encoder:
-    type: with_file_cache
-    path:
-      type: os.path.join
-      args:
-        - "@/output_dir"
-        - "word_encoder.pt"
-    config:
-      type: torchnlp.encoders.LabelEncoder
-      sample: "@/words"
-      min_occurrences: 5
-  action_sequence_encoder:
-    type: with_file_cache
-    path:
-      type: os.path.join
-      args:
-        - "@/output_dir"
-        - "action_sequence_encoder.pt"
-    config:
-      type: mlprogram.encoders.ActionSequenceEncoder
-      samples: "@/samples"
-      token_threshold: 5
-
 normalize_dataset:
   type: mlprogram.utils.transform.NormalizeGroundTruth
   normalize:
@@ -66,14 +10,54 @@ normalize_dataset:
         - - unparse
           - "@/parser.unparse"
 
+train_dataset: "@/dataset.train"
+test_dataset:
+  type: mlprogram.utils.data.transform
+  dataset: "@/dataset.test"
+  transform: "@/normalize_dataset"
+valid_dataset:
+  type: mlprogram.utils.data.transform
+  dataset: "@/dataset.valid"
+  transform: "@/normalize_dataset"
+
+encoder:
+  word_encoder:
+    type: with_file_cache
+    path:
+      type: os.path.join
+      args:
+        - "@/output_dir"
+        - "word_encoder.pt"
+    config:
+      type: torchnlp.encoders.LabelEncoder
+      sample:
+        type: mlprogram.utils.data.get_words
+        dataset: "@/train_dataset"
+        extract_reference:
+          type: mlprogram.datasets.nl2bash.TokenizeQuery
+      min_occurrences: 3
+  action_sequence_encoder:
+    type: with_file_cache
+    path:
+      type: os.path.join
+      args:
+        - "@/output_dir"
+        - "action_sequence_encoder.pt"
+    config:
+      type: mlprogram.encoders.ActionSequenceEncoder
+      samples:
+        type: mlprogram.utils.data.get_samples
+        dataset: "@/train_dataset"
+        parser: "@/parser"
+      token_threshold: 0
+
 action_sequence_reader:
   type: mlprogram.nn.nl2code.ActionSequenceReader
   num_rules: "@/encoder.action_sequence_encoder._rule_encoder.vocab_size"
   num_tokens: "@/encoder.action_sequence_encoder._token_encoder.vocab_size"
   num_node_types: "@/encoder.action_sequence_encoder._node_type_encoder.vocab_size"
-  node_type_embedding_size: 64
-  embedding_size: 128
-
+  node_type_embedding_size: "@/node_type_embedding_size"
+  embedding_size: "@/embedding_size"
 model:
   type: mlprogram.utils.share_memory
   model:
@@ -84,9 +68,9 @@ model:
         - - "encoder"
           - type: mlprogram.nn.nl2code.NLReader
             num_words: "@/encoder.word_encoder.vocab_size"
-            embedding_dim: 128
-            hidden_size: 256
-            dropout: 0.2
+            embedding_dim: "@/embedding_size"
+            hidden_size: "@/hidden_size"
+            dropout: "@/dropout"
         - - "decoder"
           - type: torch.nn.Sequential
             modules:
@@ -96,24 +80,24 @@ model:
                   - "@/action_sequence_reader"
                 - - "decoder"
                   - type: mlprogram.nn.nl2code.Decoder
-                    query_size: 256
+                    query_size: "@/hidden_size"
                     input_size:
                       type: add
                       x:
                         type: mul
                         x: 2
-                        y: 128
-                      y: 64
-                    hidden_size: 256
-                    att_hidden_size: 50
-                    dropout: 0.2
+                        y: "@/embedding_size"
+                      y: "@/node_type_embedding_size"
+                    hidden_size: "@/hidden_size"
+                    att_hidden_size: "@/attr_hidden_size"
+                    dropout: "@/dropout"
                 - - "predictor"
                   - type: mlprogram.nn.nl2code.Predictor
                     reader: "@/action_sequence_reader"
-                    embedding_size: 128
-                    query_size: 256
-                    hidden_size: 256
-                    att_hidden_size: 50
+                    embedding_size: "@/embedding_size"
+                    query_size: "@/hidden_size"
+                    hidden_size: "@/hidden_size"
+                    att_hidden_size: "@/attr_hidden_size"
 
 collate:
   type: mlprogram.utils.data.Collate
@@ -171,19 +155,19 @@ collate:
 
 synthesizer:
   type: mlprogram.synthesizers.BeamSearch
-  beam_size: 15
-  max_step_size: 100
+  beam_size: "@/beam_size"
+  max_step_size: "@/max_step_size"
   sampler:
     type: mlprogram.samplers.transform
     sampler:
       type: mlprogram.samplers.ActionSequenceSampler
       encoder: "@/encoder.action_sequence_encoder"
       is_subtype:
-        type: mlprogram.languages.python.IsSubtype
+        type: mlprogram.languages.bash.IsSubtype
       transform_input:
         type: mlprogram.utils.transform.nl2code.TransformQuery
         extract_reference:
-          type: mlprogram.datasets.django.TokenizeQuery
+          type: mlprogram.datasets.nl2bash.TokenizeQuery
         word_encoder: "@/encoder.word_encoder"
       transform_action_sequence:
         type: mlprogram.utils.transform.nl2code.TransformActionSequence
@@ -193,10 +177,8 @@ synthesizer:
       module: "@/model"
     transform: "@/parser.unparse"
 
-# Metrics
 metrics:
   accuracy:
     type: mlprogram.metrics.Accuracy
   bleu:
-    type: mlprogram.languages.python.metrics.Bleu
-
+    type: mlprogram.metrics.Bleu
