+++ configs/nl2code/nl2code_evaluate.yaml
@@ -1,0 +1,24 @@
+imports:
+  - "nl2code_base.yaml"
+
+device:
+  type: torch.device
+  type_str: "cpu"
+  index: 0
+
+output_dir: "output/output"
+
+main:
+  type: mlprogram.entrypoint.evaluate
+  workspace_dir: "output/workspace"
+  input_dir: "@/output_dir"
+  output_dir: "@/output_dir"
+  valid_dataset: "@/valid_dataset"
+  model: "@/model"
+  synthesizer: "@/synthesizer"
+  metrics: "@/metrics"
+  top_n:
+    - 1
+    - 3
+  device: "@/device"
+  n_process: 2
