+++ mlprogram/functools/__init__.py
@@ -1,2 +1,5 @@
 from mlprogram.functools.cache import file_cache  # noqa
 from mlprogram.functools.cache import FileCache  # noqa
+from mlprogram.functools.functions import Compose  # noqa
+from mlprogram.functools.functions import Sequence  # noqa
+from mlprogram.functools.functions import Map  # noqa
