+++ mlprogram/utils/__init__.py
@@ -1,4 +1,4 @@
-from mlprogram.utils.functions import Compose, Sequence, Map, Flatten, Threshold, Pick  # noqa
+from mlprogram.utils.functions import Flatten, Threshold, Pick  # noqa
 from mlprogram.utils.functions import Identity  # noqa
 from mlprogram.utils.functions import share_memory  # noqa
 from mlprogram.utils.primitives import Add, Sub, Mul, Div, IntDiv, Neg  # noqa
